+++ random/sequence.go
@@ -58,7 +58,7 @@ func (s *Sequence) Next(n uint, charsets string) string {
 		if remain == 0 {
 			cache, remain = s.src(), perDraw
 		}
-		if idx := int(cache & mask); idx < len(charsets) {
+		if idx := cache & mask; idx < uint64(len(charsets)) {
 			b[i] = charsets[idx]
 			i--
 		}
